requirements.txt
asgiref==3.4.1
astroid==2.9.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
colorama==0.4.4
colorclass==2.2.2
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-tinymce4-lite==1.8.0
docopt==0.6.2
idna==3.3
isort==5.10.1
jsmin==3.0.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
pip-upgrader==1.4.15
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
terminaltables==3.1.10
toml==0.10.2
typed-ast==1.5.1
typing_extensions==4.0.1
tzdata==2021.5
upgrade-requirements==1.7.0
urllib3==1.26.7
wincertstore==0.2
wrapt==1.13.3
